requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
django-environ==0.8.1
djangorestframework==3.12.4
environ==1.0
fsspec==2021.11.0
idna==3.3
numpy==1.21.4
pandas==1.3.4
psycopg2==2.9.2
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
